Rename components part I
    - rename all functions beside QueryPage and Core
    - create extra folder for models

Ticket: #180
