+++ src/app/Core/ExportServices/Export-Labels.service.tsx
@@ -4,8 +4,8 @@ import GetAppIcon from "@material-ui/icons/GetApp";
 import { useTranslation } from "react-i18next";
 import { secondaryColor } from "../../Shared/Style/Style-MainTheme.component";
 import { getFormattedDate } from "../getCurrentDate.service";
-import { MainFilterLabelInterface } from "../../Shared/Export.model";
-import { mainFilterAttributes } from "../../Shared/Filter.model";
+import { MainFilterLabelInterface } from "../../Shared/Model/Export.model";
+import { mainFilterAttributes } from "../../Shared/Model/Filter.model";
 
 const buttonLableStyle = (open: boolean): SerializedStyles => css`
     display: flex;
