+++ src/app/Core/ExportServices/objectsToZIP.service.ts
@@ -1,7 +1,7 @@
 import JSZip from "jszip";
 import { saveAs } from "file-saver";
-import { FilterInterface } from "../../Shared/Filter.model";
-import { ExportInterface, MainFilterLabelInterface } from "../../Shared/Export.model";
+import { FilterInterface } from "../../Shared/Model/Filter.model";
+import { ExportInterface, MainFilterLabelInterface } from "../../Shared/Model/Export.model";
 import { generateParameterHeader } from "./generateParameterHeader.service";
 import { statDataStringGenerator } from "./generateStatString.service";
 import { RAWDataStringGenerator } from "./generateRAWString.service";
