+++ src/app/Main/Body/Body-Router.component.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 import { QueryPageContainerComponent as QueryPage } from "./QueryPage/QueryPage-Container.component";
-import { HomePageLayoutComponent as HomePage } from "./HomePage/HomePage.component";
+import { HomePageLayoutComponent as HomePage } from "./HomePage/HomePage-Layout.component";
 import { LinkPageComponent as LinkPage } from "./LinkPage/LinkPage.component";
 import { InfoPageComponent as InfoPage } from "./InfoPage/InfoPage.component";
 import { DataProtectionPageComponent as DataProtection } from "./DataProtectionPage/DataProtectionPage.component";
