+++ src/app/Main/Body/LinkPage/LinkPage-LinkList.component.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { List, ListSubheader } from "@material-ui/core";
 import links from "../../../../assets/external_links.json";
-import { LinkListContentComponent as LinkListContent} from "./LinkPage-ListContent.component";
+import { LinkListContentComponent as LinkListContent} from "./LinkList-ListContent.component";
 
-export function LinkListComponent(): JSX.Element[] {
+export function LinkPageListComponent(): JSX.Element[] {
     const { t } = useTranslation(["ExternLinks"]);
     const linkJson = links;
     const elements: JSX.Element[] = [];
