+++ src/app/Shared/Model/Isolat.model.ts
