+++ src/app/Shared/Style/Style-MainTheme.component.ts
@@ -6,7 +6,7 @@ export const bfrBlue = "rgb(3, 78, 162)"; // #034ea2
 export const bfrWhite = "rgb(255, 255, 255)"; // #ffffff
 export const bfrLightgrey = "rgb(204, 204, 204)"; // #cccccc
 export const bfrBlueGrey = "rgb(184, 191, 200)"; // #b8bfc8
-export const bfrDarkBlueGrey = "rgb(51, 77, 107)";
+export const bfrDarkBlueGrey = "rgb(51, 77, 107)"; 
 export const bfrBlack = "rgb(0, 0, 0)"; // #000000
 export const bfrLightblue = "rgb(140, 190, 218)"; // #8cbeda
 export const bfrLightcyan = "rgb(209, 237, 244)"; // #d1edf4
