--- src/app/datasets.ts
@@ -1,14 +1,0 @@
-export const datasets = [
-    {
-        id: 1,
-        name: "Test Dataset 1",
-    },
-    {
-        id: 2,
-        name: "Test Dataset 2",
-    },
-    {
-        id: 3,
-        name: "Test Dataset 3",
-    },
-];
